Move angle_difference to util.geometry
